Cmm: Sink - Track hp aliasing.
See Note [Heap Kinds]. We now split heap into "new" and "old" heap
for the purpose of determining memory conflicts during CmmSink.
With an assumption that writes to memory via [Hp] don't conflict
with reads from regular heap memory.

We can turn this on/off with the new flag -fcmm-sink-alias which is now
enabled at -O2 by default.

This fixes the issues laid out in #20679.