eliminate `Refunctionalize`
This commit addresses the most salient of the issues raised during
review.  As noted in
https://gitlab.haskell.org/ghc/ghc/-/merge_requests/7442#note_409327,

> If [`Refunctionalize` is eliminated], that is only replacing
> switches on Backend constructors with switches on new datatypes'
> constructors, the MR would have been an improvement over the status
> quo.

I realize that there are a great many changes that @hsyl20 and
@Ericson2314 would like to make to the overall organization of GHC's
back end and driver.   (This is the "bottom-up approach", the most
complete explanation can be found in
https://gitlab.haskell.org/ghc/ghc/-/issues/20927#note_419064; see
also https://gitlab.haskell.org/ghc/ghc/-/issues/17957.)
I'm delighted by the bottom-up approach, and although this MR may
not advance the bottom-up agenda, I believe it is does not impede
that agenda.  And centralizing the `Backend` logic may well help
with "[splitting] the current Backend datatype to get different
types for [different] use cases" as suggested in
https://gitlab.haskell.org/ghc/ghc/-/issues/20927#note_416909.

This commit also accommodates the changes in `GHC.Tc.Gen.Foreign`,
in particular the introduction of algebraic data types to represent
errors.  The decision to represent "expected back ends" as a list of
backends, type `[Backend]`, was made in consultation with Aaron Allen,
who committed the new error-handling code.

On my local machine, `./hadrian/build -j test` passes all tests.