hadrian/tool-args: Write output to intermediate file rather than via stdout
This allows us to see the output of hadrian while it is doing the setup.

Removes the special hacks for ghc-bin and treats it as an ordinary package
instead of treating it as part of the compiler package. The extra time
needed to build the dependencies is mitigated by the fact that now
the hadrian progress messages are actually visible in stdout.
Also lets tool-args/GHCi/IDE support work for all stage0 packages in the
tree.