add Backend migration path for GHC API clients
This commit adds the module `GHC.Driver.Backend.Migration`, which
should be used only by client code, _not_ within GHC.  This module
exports type `BackendName`, which can serve as a one-for-one
replacement of the old `Backend` type in client code.  A value of
type `BackendName` is obtained by applying function `backendName`
to a value of type `Backend`.

The commit also includes a section for the user's guide.